Switch out implicit pandoc install for explicit docker container
Pandoc & LaTex for PDF generation ends up bloating the system with tons of
haskell package dependencies. Swapped to a docker container to reduce the
noise.

Bonus is that this will be easier to CI if I want to chuck it on github and try
out travis for release PDF generation.